BookjobsWebApp/requirements.txt
Django==4.0.4
mysqlclient>=2.0
django-mysql==4.5
djangorestframework==3.13.1